requirements.txt
pandas
numpy
duckdb
yfinance
requests
praw
nltk
python-dotenv
scikit-learn
pandas-datareader
